管理会员 增改 以及其他

+++ src/main/java/com/spring/model/admin/AdminUserAuthentication.java
@@ -17,7 +17,7 @@ public class AdminUserAuthentication {
 	@Pattern(regexp="(?![0-9]+$)(?![a-zA-Z]+$)[0-9A-Za-z]{6,12}",message="密码必须是6~12位数字和字母的组合")
 	private String newPassword = "";
 	
-	private int status;
+	private byte status;
 	
 	public String getUsername() {
 		return username;
@@ -37,10 +37,10 @@ public class AdminUserAuthentication {
 	public void setNewPassword(String newPassword) {
 		this.newPassword = newPassword;
 	}
-	public int getStatus() {
+	public byte getStatus() {
 		return status;
 	}
-	public void setStatus(int status) {
+	public void setStatus(byte status) {
 		this.status = status;
 	}
 }
